Do not call bash as login shell finally (#14)
This is only a hotfix which enables calls of java via non-default
locations (if its on $PATH at all). There should be a better solution
since the wrapper tries to determine a well-suiting Java version
and should finally use exactly the respective java executable.

Hint: The exec call reduces the number of processes, in particular
on Docker containers.